pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kof-notebooklm-mcp"
version = "0.1.0"
description = "MCP server for Google NotebookLM - Create notebooks, add sources, and query AI with citations"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "KeepOnFirst", email = "keeponfirst@example.com" }
]
keywords = ["mcp", "notebooklm", "google", "ai", "knowledge-management", "model-context-protocol"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "mcp>=1.0.0",
    "playwright>=1.40.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
kof-notebooklm-mcp = "kof_notebooklm_mcp.server:main"
kof-notebooklm-init = "kof_notebooklm_mcp.cli:init_auth"
kof-notebooklm-health = "kof_notebooklm_mcp.cli:check_health"
kof-notebooklm = "kof_notebooklm_mcp.cli:main_cli"

[project.urls]
Homepage = "https://github.com/keeponfirst/kof-notebooklm-mcp"
Documentation = "https://github.com/keeponfirst/kof-notebooklm-mcp#readme"
Repository = "https://github.com/keeponfirst/kof-notebooklm-mcp"
Issues = "https://github.com/keeponfirst/kof-notebooklm-mcp/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/kof_notebooklm_mcp"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
